+++ src/main/java/com/single/springboard/SpringBoardApplication.java
@@ -2,7 +2,9 @@ package com.single.springboard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
 public class SpringBoardApplication {
